F1_code_so_far
This code modelled a metric to measure a given team’s dominance in python. It produces a plot of the fraction of maximum points a team has scored over all races in a given time range.
This is everything that I have worked on so far. This code contains:
- chosen_constructors
- year_round count
- points calculations
- plots
The input validation is simple. It just checks if input name is exactly in the data, and checks if the start_year<end_year.